I added a non-monobehavior version of the INI system.
I added a non-monobehavior version of the INI system. it doesn't have complete functionality, but once it does I'll turn the old one in to a wraper for it. This will allow it to be used as a more general parser. and not just a data save-load system for the game